finished delete functionality

+++ viewrecords.php
@@ -23,7 +23,7 @@
         <td>
             <a href="view.php?id=<?php echo $r['attendee_id']?>" class="btn btn-primary">View</a>
             <a href="edit.php?id=<?php echo $r['attendee_id']?>" class="btn btn-warning">Edit</a>
-            <a href="delete.php?id=<?php echo $r['attendee_id']?>" class="btn btn-danger">Del</a>
+            <a onclick="return confirm('Confirm deletion')" href="delete.php?id=<?php echo $r['attendee_id']?>" class="btn btn-danger">Del</a>
         </td>
     </tr>
     <?php } ?>
